requirements.txt
# Core GUI Framework
PyQt6
PyQt6-sip
PyQt6-Qt6

# Async Integration for Qt
qasync

# LLM Client Libraries
ollama
google-generativeai

# --- RAG Dependencies ---
# Vector Database & Search
numpy # Required by embedding/math operations
faiss-cpu # Vector index library (CPU version)

# Embedding Model Library
sentence-transformers

# Text Splitting for RAG
langchain-text-splitters # For general text and Python code splitting

# Document Reading Libraries
PyPDF2      # For reading PDF files
python-docx # For reading DOCX files

# --- UI Enhancements & Utilities ---
Markdown    # For rendering Markdown in chat bubbles
Pillow      # For image handling (loading, resizing, encoding)

# --- Configuration & Environment ---
# Library to load environment variables from .env files
python-dotenv

# --- Development & Utility (Optional but good to have listed if used) ---
# ast # (Standard library, but good to remember if we relied on specific versions later)
# logging # (Standard library)
# asyncio # (Standard library)


qtawesome